move models to the model package
