+++ src/main/java/tacos/model/Ingredient.java
@@ -1,4 +1,4 @@
-package tacos;
+package tacos.model;
 
 import lombok.AccessLevel;
 import lombok.Data;
