+++ src/main/java/tacos/model/Taco.java
@@ -1,4 +1,4 @@
-package tacos;
+package tacos.model;
 
 import lombok.Data;
 import javax.persistence.*;
@@ -19,11 +19,16 @@ public class Taco {
     @Size(min=5, message="Name must be at least 5 characters long")
     private String name;
 
-    @ManyToMany(targetEntity = Ingredient.class)
     @NotNull(message = "You must choose at least 1 ingredient")
     @Size(min=1, message = "You must choose at least 1 ingredient")
+    @ManyToMany(targetEntity = Ingredient.class)
+    @JoinTable(name = "taco_ingredients",
+        joinColumns = @JoinColumn(name = "taco_id"),
+        inverseJoinColumns = @JoinColumn(name = "ingredient_id")
+    )
     private List<Ingredient> ingredients;
 
+    @Column(name = "created_at")
     private Date createdAt;
     @PrePersist
     void createdAt() {
